pyproject.toml
[project]
name = "enterprise-data-platform-snowflake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbt-snowflake>=1.11.3",
    "faker>=40.11.0",
    "numpy==2.4.2",
    "python-dotenv>=1.2.2",
    "snowflake-snowpark-python[pandas]>=1.47.0",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "pytest>=9.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
